requirements.txt
# Core dependencies
langchain>=0.0.267
langchain-community>=0.0.10
python-dotenv>=1.0.0

# AWS Bedrock dependencies
boto3>=1.28.0
botocore>=1.31.0
langchain-aws>=0.1.0

# Oracle Database dependencies
oracledb>=1.4.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Testing
pytest>=7.3.1
